# Add some caching for the yaml provider