Low hanging fruit for yaml perf.

Bench diffs before and after:

```
$ benchcmp old.txt new.txt
benchmark                                            old ns/op     new ns/op     delta
BenchmarkYAMLCreateSingleFile-8                      50558         52130         +3.11%
BenchmarkYAMLCreateMultiFile-8                       79316         81416         +2.65%
BenchmarkYAMLSimpleGetLevel1-8                       224           29.5          -86.83%
BenchmarkYAMLSimpleGetLevel3-8                       470           29.5          -93.72%
BenchmarkYAMLSimpleGetLevel7-8                       925           29.3          -96.83%
BenchmarkYAMLPopulateStruct-8                        2752          978           -64.46%
BenchmarkYAMLPopulateStructNested-8                  5467          3039          -44.41%
BenchmarkYAMLPopulateStructNestedMultipleFiles-8     9918          4134          -58.32%

benchmark                                            old allocs     new allocs     delta
BenchmarkYAMLCreateSingleFile-8                      120            121            +0.83%
BenchmarkYAMLCreateMultiFile-8                       206            207            +0.49%
BenchmarkYAMLSimpleGetLevel1-8                       4              0              -100.00%
BenchmarkYAMLSimpleGetLevel3-8                       8              0              -100.00%
BenchmarkYAMLSimpleGetLevel7-8                       16             0              -100.00%
BenchmarkYAMLPopulateStruct-8                        32             12             -62.50%
BenchmarkYAMLPopulateStructNested-8                  68             39             -42.65%
BenchmarkYAMLPopulateStructNestedMultipleFiles-8     126            52             -58.73%

benchmark                                            old bytes     new bytes     delta
BenchmarkYAMLCreateSingleFile-8                      10784         10832         +0.45%
BenchmarkYAMLCreateMultiFile-8                       19792         19840         +0.24%
BenchmarkYAMLSimpleGetLevel1-8                       64            0             -100.00%
BenchmarkYAMLSimpleGetLevel3-8                       128           0             -100.00%
BenchmarkYAMLSimpleGetLevel7-8                       256           0             -100.00%
BenchmarkYAMLPopulateStruct-8                        720           256           -64.44%
BenchmarkYAMLPopulateStructNested-8                  1408          736           -47.73%
BenchmarkYAMLPopulateStructNestedMultipleFiles-8     2656          952           -64.16%
```