@glibsm 2016-11-07
@uber-go/service-framework