@shawnburke 2016-11-11
Suggestion: Have a benchmark for a situation where some values from from Yaml and some come from env.